Move tie commands from Score into EditTie class
Move cmdAddTie, cmdToggleTie, cmdToggleLaissezVib and cmdTieNoteList out
of Score into a new EditTie static class in editing/edittie.h, and update
all call sites.
